requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.1.5
django-bootstrap4==2.3.1
django-crispy-forms==1.10.0
django-filter==2.4.0
django-tables2==2.3.4
importlib-metadata==2.1.1
numpy==1.19.5
pandas==1.2.1
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
zipp==3.4.0
